Cleanup and add overload for ListRemoteReferences
